LayoutTests/http/tests/navigation/page-cache-getUserMedia-pending-promise-expected.txt

Tests that a page with a pending getUserMedia request is able to go into the back/forward cache.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


pageshow - not from cache
pagehide - entering cache
pageshow - from cache
PASS Page was restored from Page Cache
PASS Promised was resolved
PASS restoredFromPageCache is true
PASS successfullyParsed is true

TEST COMPLETE
